Replacing float primitive number with rug::Float

+++ backend/libs/exchanges/observers/src/binance/pubsub.rs
@@ -1,5 +1,7 @@
+use ::rug::Float;
+
 use ::subscribe::pubsub;
 
 use super::entities::BookTicker;
 
-pubsub!(pub, BookTickerPubSub, BookTicker<f64>, "BinanceTrades",);
+pubsub!(pub, BookTickerPubSub, BookTicker<Float>, "BinanceTrades",);
